add support for castem discretization in 1D

+++ src/System/getFunction.c
@@ -53,7 +53,7 @@ int tfel_getBool(void *lib,
   if(n==0){
     return -1;
   }  
-  return *n;
+  return (int) *n;
 }
 
 char **
